requirements-render.txt
# CIL-BACKEND on Render: same as requirements.txt but WITHOUT streamlit/plotly
# to stay under 512MB. Tariff router is disabled on Render (see main.py).
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic[email]>=2.0.0
email-validator>=2.0.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
duckdb>=0.9.0
reportlab>=4.1.0
anthropic>=0.28.0
aiofiles>=23.2.0
httpx>=0.27.0
polars>=0.19.0
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0
